requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
googlemaps==4.10.0
SQLAlchemy==2.0.28
Werkzeug==3.0.1 